electron/backend/requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
distlib==0.3.7
docutils==0.20.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230314
pandas==2.0.3
pexpect==4.8.0
platformdirs==3.9.1
ptyprocess==0.7.0
pyarrow==12.0.1
Pygments==2.15.1
pylance==0.5.9
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
rake-nltk==1.0.6
regex==2023.6.3
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.1
sh==2.0.4
six==1.16.0
sklearn==0.0.post7
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.24.0
